todo detect and save auto configuration
Later parts of the codebase expect .PDK or .PUPPET, so the .AUTO is breaking
things like the TCPIP handler (empty command)

So in the aggregateconfig class, detect which we'll use and the munge the
installType setting to relfect that.

Updates the error message handling so it flips based on the original installType
setting.